# Promql: reuse LabelBuilder in aggregations